Add files via upload
This is a very simple java program i made for my internship. It is a library system where we can add books, issue and return them. I used classes and object concept in this project. The code runs in the console and i have tried to make it easy and basic so that i can understand how java works. I am still learning java so the code is not advanced.